requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
google-generativeai>=0.3.2
